updating elastic correctly
